fix: update karyawan

+++ app/Http/Controllers/Location/DistrictController.php
@@ -9,6 +9,7 @@ use App\Http\Resources\Location\DistrictResource;
 use App\Models\Location\District;
 use App\Models\Location\Regency;
 use App\Traits\RegionTrait;
+use Illuminate\Http\Exceptions\ThrottleRequestsException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
@@ -32,14 +33,14 @@ class DistrictController extends Controller
 
         $districtResource = DistrictResource::collection($districts);
 
-        $component = $request->path().'/index';
+        $component = $request->path() . '/index';
 
         return inertia($component, [
             'page_settings' => [
                 'title' => 'Kecamatan',
             ],
-            'regencies' => fn () => $regencies,
-            'districts' => fn () => $districtResource,
+            'regencies' => fn() => $regencies,
+            'districts' => fn() => $districtResource,
         ]);
     }
 
@@ -75,18 +76,18 @@ class DistrictController extends Controller
                     'name' => $reqValidated['name'],
                 ]);
 
-            if (! $regency) {
+            if (!$regency) {
                 flashMessage('Gagal Menambahkan Kecamatan', 'Kecamatan sudah ada', 'error');
 
                 return redirect()->back()->withErrors($validatedData->errors());
             }
 
-            DB::commit();
             flashMessage('Kecamatan Ditambahkan', 'Kecamatan berhasil ditambahkan');
+            DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
             flashMessage('Gagal Menambahkan Kecamatan', 'Terjadi kesalahan saat menambahkan kecamatan', 'error');
-            Log::error('Kecamatan Store: '.json_encode($th->getMessage(), JSON_PRETTY_PRINT));
+            Log::error('Kecamatan Store: ' . json_encode($th->getMessage(), JSON_PRETTY_PRINT));
         } finally {
             return redirect()->back();
         }
@@ -108,12 +109,12 @@ class DistrictController extends Controller
 
             $district->update($request->only('regency_id', 'code', 'name'));
 
-            DB::commit();
             flashMessage('Kecamatan Diperbarui', 'Kecamatan berhasil diperbarui');
+            DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
             flashMessage('Gagal Memperbarui Kecamatan', 'Terjadi kesalahan saat memperbarui kecamatan', 'error');
-            Log::error('Kecamatan Update: '.json_encode($th->getMessage(), JSON_PRETTY_PRINT));
+            Log::error('Kecamatan Update: ' . json_encode($th->getMessage(), JSON_PRETTY_PRINT));
         } finally {
             return redirect()->back();
         }
@@ -127,14 +128,18 @@ class DistrictController extends Controller
         try {
             DB::beginTransaction();
 
-            $district->delete();
+            if ($district->exists) {
+                $district->delete();
+            } else {
+                throw new ThrottleRequestsException('Kantor Cabang tidak ditemukan');
+            }
 
-            DB::commit();
             flashMessage('Kecamatan Dihapus', 'Kecamatan berhasil dihapus');
+            DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
             flashMessage('Gagal Menghapus Kecamatan', 'Terjadi kesalahan saat menghapus kecamatan', 'error');
-            Log::error('Kecamatan Delete: '.json_encode($th->getMessage(), JSON_PRETTY_PRINT));
+            Log::error('Kecamatan Delete: ' . json_encode($th->getMessage(), JSON_PRETTY_PRINT));
         } finally {
             return redirect()->back();
         }
@@ -160,7 +165,7 @@ class DistrictController extends Controller
             // Ambil hasil dari permintaan
             $regency = Regency::query()
                 ->where('code', $request->get('code'))->first();
-            $districts = (object) $responses[0]->json();
+            $districts = (object)$responses[0]->json();
 
             foreach ($districts->value as $district) {
                 District::query()
@@ -173,12 +178,12 @@ class DistrictController extends Controller
                     ]);
             }
 
-            DB::commit();
             flashMessage('Kecamatan Disinkronkan', 'Kecamatan berhasil disinkronkan');
+            DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
             flashMessage('Gagal Menyinkronkan Kecamatan', json_encode($th->getMessage(), JSON_PRETTY_PRINT), 'error');
-            Log::error('Kecamatan Sync: '.json_encode($th->getMessage(), JSON_PRETTY_PRINT));
+            Log::error('Kecamatan Sync: ' . json_encode($th->getMessage(), JSON_PRETTY_PRINT));
         } finally {
             return redirect()->back();
         }
