+++ app/Http/Controllers/Location/ProvinceController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\Location\ProvinceResource;
 use App\Models\Location\Province;
 use App\Traits\RegionTrait;
+use Illuminate\Http\Exceptions\ThrottleRequestsException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
@@ -57,8 +58,8 @@ class ProvinceController extends Controller
             DB::beginTransaction();
             Province::query()
                 ->create($request->only('code', 'name'));
-            DB::commit();
             flashMessage('Provinsi Ditambahkan', 'Provinsi berhasil ditambahkan');
+            DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
             flashMessage('Gagal Menambahkan Provinsi', 'Terjadi kesalahan saat menambahkan provinsi', 'error');
@@ -90,8 +91,8 @@ class ProvinceController extends Controller
 
             $province->update($request->only('code', 'name'));
 
-            DB::commit();
             flashMessage('Provinsi Diperbarui', 'Provinsi berhasil diperbarui');
+            DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
             flashMessage('Gagal Memperbarui Provinsi', 'Terjadi kesalahan saat memperbarui provinsi', 'error');
@@ -120,10 +121,15 @@ class ProvinceController extends Controller
 
                 return redirect()->route('province.index');
             }
-            $province->delete();
 
-            DB::commit();
+            if($province->exists) {
+                $province->delete();
+            } else {
+                throw new ThrottleRequestsException('Provinsi tidak ditemukan');
+            }
+
             flashMessage('Provinsi Dihapus', 'Provinsi berhasil dihapus');
+            DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
             flashMessage('Gagal Menghapus Provinsi', 'Terjadi kesalahan saat menghapus provinsi', 'error');
@@ -153,8 +159,8 @@ class ProvinceController extends Controller
                 ]);
             }
 
-            DB::commit();
             flashMessage('Provinsi Tersinkron', 'Provinsi berhasil disinkronisasi');
+            DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
             flashMessage('Gagal Menyinkronkan Provinsi', 'Terjadi kesalahan saat menyinkronkan provinsi', 'error');
