+++ app/Http/Controllers/Location/RegencyController.php
@@ -7,6 +7,7 @@ use App\Http\Resources\Location\RegencyResource;
 use App\Models\Location\Province;
 use App\Models\Location\Regency;
 use App\Traits\RegionTrait;
+use Illuminate\Http\Exceptions\ThrottleRequestsException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
@@ -29,14 +30,14 @@ class RegencyController extends Controller
 
         $regenciesResource = RegencyResource::collection($regencies);
 
-        $component = $request->path().'/index';
+        $component = $request->path() . '/index';
 
         return inertia($component, [
             'page_settings' => [
                 'title' => 'Kabupaten',
             ],
-            'provinces' => fn () => $provinces,
-            'regencies' => fn () => $regenciesResource,
+            'provinces' => fn() => $provinces,
+            'regencies' => fn() => $regenciesResource,
         ]);
     }
 
@@ -73,12 +74,12 @@ class RegencyController extends Controller
                     'name' => $request->get('name'),
                 ]);
 
-            DB::commit();
             flashMessage('Kabupaten Ditambahkan', 'Kabupaten berhasil ditambahkan');
+            DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
             flashMessage('Gagal Menambahkan Kabupaten', 'Terjadi kesalahan saat menambahkan kabupaten', 'error');
-            Log::error('Kabupaten Store: '.json_encode($th->getMessage(), JSON_PRETTY_PRINT));
+            Log::error('Kabupaten Store: ' . json_encode($th->getMessage(), JSON_PRETTY_PRINT));
         } finally {
             return redirect()->route('regency.index');
         }
@@ -92,8 +93,8 @@ class RegencyController extends Controller
     {
         $request->validate([
             'province_id' => 'required|exists:provinces,id',
-            'code' => 'required|string|unique:regencies,code,'.$regency->id,
-            'name' => 'required|string|unique:regencies,name,'.$regency->id,
+            'code' => 'required|string|unique:regencies,code,' . $regency->id,
+            'name' => 'required|string|unique:regencies,name,' . $regency->id,
         ], [
             'province_id.required' => 'Provinsi wajib diisi',
             'province_id.exists' => 'Provinsi tidak ditemukan',
@@ -110,12 +111,12 @@ class RegencyController extends Controller
 
             $regency->update($request->only('province_id', 'code', 'name'));
 
-            DB::commit();
             flashMessage('Kabupaten Diperbarui', 'Kabupaten berhasil diperbarui');
+            DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
             flashMessage('Gagal Memperbarui Kabupaten', 'Terjadi kesalahan saat memperbarui kabupaten', 'error');
-            Log::error('Kabupaten Update: '.json_encode($th->getMessage(), JSON_PRETTY_PRINT));
+            Log::error('Kabupaten Update: ' . json_encode($th->getMessage(), JSON_PRETTY_PRINT));
         } finally {
             return redirect()->route('regency.index');
         }
@@ -129,14 +130,18 @@ class RegencyController extends Controller
         try {
             DB::beginTransaction();
 
-            $regency->delete();
+            if ($regency->exists) {
+                $regency->delete();
+            } else {
+                throw new ThrottleRequestsException('Kabupaten tidak ditemukan');
+            }
 
-            DB::commit();
             flashMessage('Kabupaten Dihapus', 'Kabupaten berhasil dihapus');
+            DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
             flashMessage('Gagal Menghapus Kabupaten', 'Terjadi kesalahan saat menghapus kabupaten', 'error');
-            Log::error('Kabupaten Delete: '.json_encode($th->getMessage(), JSON_PRETTY_PRINT));
+            Log::error('Kabupaten Delete: ' . json_encode($th->getMessage(), JSON_PRETTY_PRINT));
         } finally {
             return redirect()->route('regency.index');
         }
@@ -162,7 +167,7 @@ class RegencyController extends Controller
             // Ambil hasil dari permintaan
             $province = Province::query()
                 ->where('code', $request->get('code'))->first();
-            $regencies = (object) $responses[0]->json();
+            $regencies = (object)$responses[0]->json();
 
             foreach ($regencies->value as $regency) {
                 Regency::query()
@@ -175,12 +180,12 @@ class RegencyController extends Controller
                     ]);
             }
 
-            DB::commit();
             flashMessage('Kabupaten Disinkronkan', 'Kabupaten berhasil disinkronkan');
+            DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
             flashMessage('Gagal Menyinkronkan Kabupaten', json_encode($th->getMessage(), JSON_PRETTY_PRINT), 'error');
-            Log::error('Kabupaten Sync: '.json_encode($th->getMessage(), JSON_PRETTY_PRINT));
+            Log::error('Kabupaten Sync: ' . json_encode($th->getMessage(), JSON_PRETTY_PRINT));
         } finally {
             return redirect()->route('regency.index');
         }
